remove common package and move implementation to portal package
