--- common/utils.go
@@ -1,42 +1,0 @@
-// This file is part of MinIO Kubernetes Cloud
-// Copyright (c) 2019 MinIO, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package common
-
-import (
-	"crypto/rand"
-	"crypto/sha256"
-	"io"
-
-	"encoding/base64"
-	"fmt"
-)
-
-// GetRandString generates a random string with the defined size length
-func GetRandString(size int, method string) (string, error) {
-	rb := make([]byte, size)
-	if _, err := io.ReadFull(rand.Reader, rb); err != nil {
-		return "", err
-	}
-
-	randStr := base64.URLEncoding.EncodeToString(rb)
-	if method == "sha256" {
-		h := sha256.New()
-		h.Write([]byte(randStr))
-		randStr = fmt.Sprintf("%x", h.Sum(nil))
-	}
-	return randStr, nil
-}
